requirements.txt
dashscope==1.25.0
ipywidgets==8.1.8
langchain==1.2.15
langchain-community==0.4.1
langchain-litellm==0.6.4
langchain-huggingface==1.0.1
langchain-qdrant==1.1.0
langchain-mcp-adapters==0.1.14
langchain-tavily==0.2.17
langgraph-checkpoint-postgres==3.0.1
langgraph==1.1.6
langsmith==0.4.59
pypdf==6.2.0
python-dotenv==1.2.1
psycopg-binary==3.2.13
rich==14.2.0
sentence-transformers==5.1.2